Move MaintenanceController related code to its own package
